Update Docker container to Python 3.10, add and update dependencies
The codebase has references to some features that were introduced in
3.10 and was causing the old Docker instance to crash due to syntax
error.

In addition, with the updated version, the version of `setuptools` in
`requirements.txt` was marked as conflicting due to packages requiring
higher versions when on 3.10. This commit does not update other
dependencies, but this file ideally should be cleaned up to only
mention any top-level dependencies.