Reimplemented business logic for `/event remove`
Replaces codepath for deleting partially deleted events (i.e. events
that do not exist in the database but have a server role) with a message
saying to delete the orphan server role manually. The reason why we
choose to *not* delete the role via the bot is because the bot can
potentially delete other roles not related to any event by accident.