chore: Add beanie ODM (#1)
* Added Beanie ODM initalizations

* Migrated cron collection to use ODM

* Created models.py file

* Changed which fields are indexable in `Cron` model

* Refactored remaining refs to cron collection DB operations

* Migrated ping collection to ODM

* Migrated tag collection to ODM

* Migrated invitational collection to ODM

* Migrated event collection to ODM

* Reimplemented business logic for `/event remove`

Replaces codepath for deleting partially deleted events (i.e. events
that do not exist in the database but have a server role) with a message
saying to delete the orphan server role manually. The reason why we
choose to *not* delete the role via the bot is because the bot can
potentially delete other roles not related to any event by accident.

* Migrated censor collection to ODM

* Migrated settings collection to ODM

* Remove `REPORTS` global

This was an unused global that used to be used for keeping track of
report discord message IDs. Users could submit reports using `/report`
and staff members could react to report messages to take action.

See commits for the last read/write references to the global:
- d61750b61d8091b82ffdaef87d01fa92237b8e94
- 191454883e1f319f5651303a2e7b9815ab62bf7c
- 1f4c01e09fb067e673f7e3901b26f7955716c4aa

* Remove redundant assert statements

* Remove mongo database wrapper class

At this point, as no functions within the class were being used, so it
is safe to remove. From now on, all database interactions should use
Beanie.

* Remove `get_reports()` and unused helper functions

* Add migrations folder