feat: Add top-level sync command (#5)
* Add top-level sync command and sync script

It is generally best practice to not rate limit the slash command
registration endpoint, so syncing should be done manually rather than on
ready. A script is provided to simply login the bot and sync commands in
case the sync command signature ever gets modified.

* Remove sync_commands.py script

Syncing commands now should be done using the legacy command

* Change argument from bool to string literal

* Add error handler and help text for sync command

* Remove separate sync command name variable

* Fix sync command check from allowing messages that start with "sync"

* Fix prefix in command regex to be contructed from `BOT_PREFIX`

* Add 60 second cooldown per guild for sync command