feat: `/event` command group rework (#8)
* feat: Add event role commands

These allow for event roles to be enabled and disabled

* chore: Disable `/event remove`

The current behavior conflicts with the role disable commands and staff
might confuse this command with the role toggle commands. Best to just
disable this with a message and then figure a better implementation
later.

* feat: Add event role batch commands

* fix: Add failure logic to `/events`

* fix: Fix autocomplete to only show events with enabled roles

* chore: Use `interaction.followup` for `/event` subsequent response messages

* feat: Add support to optionally not enable role on event creation

* chore: Add typing indicator for batch role command

* chore: Remove `/events`

This command is now obsolete due to the onboarding survey. With the
survey, the user's choices persist even when the user leaves, so there
is no need for them to re-add their roles upon rejoin.

Please note that during rollover, event roles need to be enabled and the
survey needs to be updated with said roles accordingly.